Dev2

+++ cilantro/storage/sqldb.py
@@ -1,21 +1,19 @@
-import mysql, mysql.connector, cilantro, os
+import mysql.connector, cilantro, os
 from configparser import SafeConfigParser
 
 path = os.path.dirname(cilantro.__path__[0])
 config = SafeConfigParser()
 config.read('{}/db_conf.ini'.format(path))
 
+
 class SQLDB():
-    connection = mysql.connector.connect(
-        host=config.get('DB','hostname'),
-        user=config.get('DB','username'),
-        passwd=config.get('DB','password'),
-        charset='binary'
-    )
-    cursor = connection.cursor()
+    connection = None
+    cursor = None
     database = config.get('DB','database')
     def __enter__(self, database=None, reset=False):
         database = database or self.database
+        if self.connection == None:
+            self.force_start()
         if reset:
             self.drop_db(database)
             self.setup_db(database)
@@ -24,6 +22,7 @@ class SQLDB():
         return self.connection, self.cursor
     def __exit__(self, type, value, traceback):
         return False
+
     @classmethod
     def force_start(cls):
         cls.connection = mysql.connector.connect(
@@ -34,8 +33,11 @@ class SQLDB():
         )
         cls.cursor = cls.connection.cursor()
         cls.setup_db()
+
     @classmethod
     def setup_db(cls, database=None):
+        if cls.connection == None:
+            cls.force_start()
         database = database or cls.database
         cls.cursor.execute("""
             CREATE DATABASE IF NOT EXISTS {}
@@ -44,21 +46,26 @@ class SQLDB():
             USE {}
             """.format(database))
         cls.build_tables()
+
     @classmethod
     def drop_db(cls, database=None):
-        cls.cursor.execute("""
-            DROP DATABASE IF EXISTS {}
-            """.format(database or cls.database))
+        if cls.cursor:
+            cls.cursor.execute("""
+                DROP DATABASE IF EXISTS {}
+                """.format(database or cls.database))
+
     @classmethod
     def reset_db(cls, database=None):
         cls.drop_db(database)
         cls.setup_db(database)
+
     @classmethod
     def truncate_tables(cls, *tables):
         for table in tables:
             cls.cursor.execute("""
                 TRUNCATE TABLE {}
             """.format(table))
+
     @classmethod
     def build_tables(cls):
         cls.cursor.execute("""
@@ -89,15 +96,3 @@ class SQLDB():
                 state BLOB NOT NULL
             )
         """)
-
-if __name__ == '__main__':
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-d', '--drop', action='store_true', default=False)
-    parser.add_argument('-b', '--build-tables', action='store_true', default=False)
-    args = parser.parse_args()
-    if args.drop:
-        SQLDB.drop_db()
-    SQLDB.setup_db()
-else:
-    SQLDB.setup_db()
